refactor: move reference marker manipulation funciton to top level of lib
